Add Job Declarator Client struct
With the goal to write more tests, this commit moves JDC initialisation
logic from `main.rs` to `lib.rs` so its easier to test and also separate
the actual `lib` code from the binary.
